feat: mirg docs form packages to diocs
