+++ docs/.vitepress/cache/deps/_metadata.json
@@ -1,25 +1,25 @@
 {
-  "hash": "e4788165",
-  "configHash": "c7c548f0",
+  "hash": "edc7a408",
+  "configHash": "4231b946",
   "lockfileHash": "f8ebd001",
-  "browserHash": "46dc6570",
+  "browserHash": "f38cf3fb",
   "optimized": {
     "vue": {
       "src": "../../../../node_modules/.pnpm/vue@3.4.38_typescript@4.9.5/node_modules/vue/dist/vue.runtime.esm-bundler.js",
       "file": "vue.js",
-      "fileHash": "213cf8f7",
+      "fileHash": "82a53974",
       "needsInterop": false
     },
     "vitepress > @vue/devtools-api": {
       "src": "../../../../node_modules/.pnpm/@vue+devtools-api@7.3.8/node_modules/@vue/devtools-api/dist/index.js",
       "file": "vitepress___@vue_devtools-api.js",
-      "fileHash": "33623f1d",
+      "fileHash": "1e413fb1",
       "needsInterop": false
     },
     "vitepress > @vueuse/core": {
       "src": "../../../../node_modules/.pnpm/@vueuse+core@11.0.1_vue@3.4.38_typescript@4.9.5_/node_modules/@vueuse/core/index.mjs",
       "file": "vitepress___@vueuse_core.js",
-      "fileHash": "8fda268c",
+      "fileHash": "d97ce564",
       "needsInterop": false
     }
   },
