+++ docs/zh/guide/advanced-custom-text.md
@@ -1,0 +1,18 @@
+# Custom Notification Text
+
+```ts
+// vite.config.ts
+export default defineConfig({
+  plugins: [
+    vue(),
+    webUpdateNotice({
+      notificationProps: {
+        title: "system update",
+        description: "System update, please refresh the page",
+        buttonText: "refresh",
+      },
+    }),
+  ],
+});
+```
+
