--- packages/.vitepress/config/en.ts
@@ -1,120 +1,0 @@
-import { createRequire } from 'module'
-import { type DefaultTheme, defineConfig } from 'vitepress'
-
-const require = createRequire(import.meta.url)
-const pkg = require('vitepress/package.json')
-
-export const en = defineConfig({
-  lang: 'en-US',
-  description: 'Detect webpage updates and notify user to reload. support vite, umijs and webpack.',
-
-  themeConfig: {
-    nav: nav(),
-
-    sidebar: {
-      '/guide/': { base: '/guide/', items: sidebarGuide() },
-      '/reference/': { base: '/reference/', items: sidebarReference() },
-    },
-
-    editLink: {
-      pattern: 'https://github.com/vuejs/vitepress/edit/main/docs/:path',
-      text: 'Edit this page on GitHub',
-    },
-
-    footer: {
-      message: 'Released under the MIT License.',
-      copyright: 'Copyright © 2019-present ArthurDarkstone',
-    },
-  },
-})
-
-function nav(): DefaultTheme.NavItem[] {
-  return [
-    {
-      text: 'Guide',
-      link: '/guide/what-is-vitepress',
-      activeMatch: '/guide/',
-    },
-    {
-      text: 'Reference',
-      link: '/reference/site-config',
-      activeMatch: '/reference/',
-    },
-    {
-      text: pkg.version,
-      items: [
-        {
-          text: 'Changelog',
-          link: 'https://github.com/vuejs/vitepress/blob/main/CHANGELOG.md',
-        },
-        {
-          text: 'Contributing',
-          link: 'https://github.com/vuejs/vitepress/blob/main/.github/contributing.md',
-        },
-      ],
-    },
-  ]
-}
-
-function sidebarGuide(): DefaultTheme.SidebarItem[] {
-  return [
-    {
-      text: 'Introduction',
-      collapsed: false,
-      items: [
-        { text: 'What\'s Web Update Notify?', link: 'what-is-wun' },
-        { text: 'Getting Started', link: 'getting-started' },
-        { text: 'Deploy', link: 'deploy' },
-      ],
-    },
-
-    {
-      text: 'Introduction',
-      collapsed: false,
-      items: [
-        { text: 'What\'s Web Update Notify?', link: 'what-is-wun' },
-        { text: 'Getting Started', link: 'getting-started' },
-        { text: 'Deploy', link: 'deploy' },
-      ],
-    },
-
-    {
-      text: 'Config & API Reference',
-      base: '/reference/',
-      link: 'site-config',
-    },
-  ]
-}
-
-function sidebarReference(): DefaultTheme.SidebarItem[] {
-  return [
-    {
-      text: 'Reference',
-      items: [
-        { text: 'Site Config', link: 'site-config' },
-        { text: 'Frontmatter Config', link: 'frontmatter-config' },
-        { text: 'Runtime API', link: 'runtime-api' },
-        { text: 'CLI', link: 'cli' },
-        {
-          text: 'Default Theme',
-          base: '/reference/default-theme-',
-          items: [
-            { text: 'Overview', link: 'config' },
-            { text: 'Nav', link: 'nav' },
-            { text: 'Sidebar', link: 'sidebar' },
-            { text: 'Home Page', link: 'home-page' },
-            { text: 'Footer', link: 'footer' },
-            { text: 'Layout', link: 'layout' },
-            { text: 'Badge', link: 'badge' },
-            { text: 'Team Page', link: 'team-page' },
-            { text: 'Prev / Next Links', link: 'prev-next-links' },
-            { text: 'Edit Link', link: 'edit-link' },
-            { text: 'Last Updated Timestamp', link: 'last-updated' },
-            { text: 'Search', link: 'search' },
-            { text: 'Carbon Ads', link: 'carbon-ads' },
-          ],
-        },
-      ],
-    },
-  ]
-}
