--- packages/en/guide/what-is-vitepress.md
@@ -1,5 +1,0 @@
-# W
-
-23
-2
-23
